pyproject.toml
[tool.poetry]
name = "sapp"
version = "0.1.0"
description = ""
authors = ["YewtsingMu <mooyewtsing@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.62.0"
numpy = "^1.19.4"
faiss-cpu = "^1.6.5"
transformers = "^4.0.0"
onnxruntime = "^1.5.2"
uvicorn = "^0.13.0"
psutil = "^5.7.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
